requirements.txt
Django >= 1.9.5
python_social_auth >= 0.2.14
djangorestframework >= 3.3.3
social >= 0.1.dev3
